+++ lib/ua_inspector/result/client.ex
@@ -4,16 +4,18 @@ defmodule UAInspector.Result.Client do
   """
 
   @type t :: %__MODULE__{
-    engine:  String.t | :unknown,
-    name:    String.t | :unknown,
-    type:    String.t | :unknown,
-    version: String.t | :unknown
+    engine:         String.t | :unknown,
+    engine_version: String.t | :unknown,
+    name:           String.t | :unknown,
+    type:           String.t | :unknown,
+    version:        String.t | :unknown
   }
 
   defstruct [
-    engine:  :unknown,
-    name:    :unknown,
-    type:    :unknown,
-    version: :unknown
+    engine:         :unknown,
+    engine_version: :unknown,
+    name:           :unknown,
+    type:           :unknown,
+    version:        :unknown
   ]
 end