+++ test/ua_inspector/parser_test.exs
@@ -42,7 +42,7 @@ defmodule UAInspector.ParserTest do
     agent  = "Mozilla/5.0 (iPad; CPU OS 7_0_4 like Mac OS X) AppleWebKit/537.51.1 (KHTML, like Gecko) Version/7.0 Mobile/11B554a Safari/9537.53"
     parsed = %Result{
       user_agent: agent,
-      client:     %Result.Client{ engine: "WebKit", name: "Mobile Safari", type: "browser", version: "7.0" },
+      client:     %Result.Client{ engine: "WebKit", engine_version: "537.51.1", name: "Mobile Safari", type: "browser", version: "7.0" },
       device:     %Result.Device{ brand: "Apple", model: "iPad", type: "tablet" },
       os:         %Result.OS{ name: "iOS", version: "7.0.4" }
     }